% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi_square2_test.R
\name{Asymp_chi2}
\alias{Asymp_chi2}
\title{Chi-square2 Test}
\usage{
Asymp_chi2(A, B, sig)
}
\arguments{
\item{A}{cell array containing networks in 1st population; each cell is a sparse adjacency matrix}

\item{B}{cell array containing networks in 2nd population, each cell being a sparse adjacency matrix}

\item{sig}{significance level for acceptance of null hypothesis}
}
\value{
acceptance/rejection decision & p-value for a normal distribution
}
\description{
The Asymp-Chi2 test is an asymptotic test for two-sample testing of large graphs.
The test is used to determine whether two graphs (or graph populations) have the same population adjacency.
The Asymp-Chi2 test involves computing a test statistic T_chi2 (which is a sum of squared differences between the sample mean differences and their estimated variances) \cr\cr
The test involves computing rejection decision and the p-value, which is the probability that the null hypothesis (i.e., both graphs have the same population adjacency) is true.
}
\examples{
simu = twosamp_twoblock(n = 4, m = 4, p = 0.3, q = 0.3, epsilon = 0.7)
test_result = Asymp_chi2(simu$A_G, simu$A_H, 0.05)

}
