% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TW_test.R
\name{Asymp_TW}
\alias{Asymp_TW}
\title{Tracy Widom Test}
\usage{
Asymp_TW(A, B, sig, r)
}
\arguments{
\item{A}{cell array containing networks in 1st population; each cell is a sparse adjacency matrix}

\item{B}{cell array containing networks in 2nd population, each cell being a sparse adjacency matrix}

\item{sig}{significance level for acceptance of null hypothesis}

\item{r}{r communities in G}
}
\value{
acceptance/rejection decision & p-value for a normal distribution
}
\description{
Tracy Widom Test
}
\details{
The TW test is a two-sample testing procedure for large graphs.
The TW test is based on the Wishart distribution and uses the Tracy-Widom law to approximate the null distribution of a test statistic.\cr\cr
The TW test then uses the Tracy-Widom law to approximate the null distribution of T_TW under certain assumptions about the underlying distributions.
If the value of T_TW exceeds a certain threshold, then the null hypothesis (i.e., both populations have the same population adjacency) is rejected.
}
\examples{
simu = twosamp_twoblock(n = 4, m = 4, p = 0.3, q = 0.3, epsilon = 0.7)
test_result = Asymp_TW(simu$A_G, simu$A_H, 0.05, 100)

}
