% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Summaries.R
\name{compute_all_F}
\alias{compute_all_F}
\title{Compute all F1–F9 for a single adjacency matrix}
\usage{
compute_all_F(adj)
}
\arguments{
\item{adj}{Adjacency matrix}
}
\value{
Named numeric vector with F1..F9
}
\description{
Compute all F1–F9 for a single adjacency matrix
}
