% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Summaries.R
\name{compute_F1}
\alias{compute_F1}
\title{Compute F1: Number of edges in the graph}
\usage{
compute_F1(adj)
}
\arguments{
\item{adj}{Adjacency matrix (numeric, symmetric, binary)}
}
\value{
Total number of edges (numeric)
}
\description{
Compute F1: Number of edges in the graph
}
