% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Summaries.R
\name{compute_F2}
\alias{compute_F2}
\title{Compute F2: Maximum degree}
\usage{
compute_F2(adj)
}
\arguments{
\item{adj}{Adjacency matrix (numeric, symmetric, binary)}
}
\value{
Maximum node degree (numeric)
}
\description{
Compute F2: Maximum degree
}
