% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Summaries.R
\name{compute_F3}
\alias{compute_F3}
\title{Compute F3: Largest eigenvalue of the adjacency matrix}
\usage{
compute_F3(adj)
}
\arguments{
\item{adj}{Adjacency matrix (numeric, symmetric)}
}
\value{
Largest eigenvalue (numeric), as proxy for max average degree
}
\description{
Compute F3: Largest eigenvalue of the adjacency matrix
}
