% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Summaries.R
\name{compute_F4}
\alias{compute_F4}
\alias{compute_F5}
\alias{compute_F6}
\title{Compute F4-6: Scan statistics (1,2,3-hop neighborhood)}
\usage{
compute_F4(adj)

compute_F5(adj)

compute_F6(adj)
}
\arguments{
\item{adj}{Adjacency matrix}
}
\value{
Maximum number of edges in a neighbourhood of a given size
}
\description{
Compute F4-6: Scan statistics (1,2,3-hop neighborhood)
}
