% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Summaries.R
\name{compute_F7}
\alias{compute_F7}
\title{Compute F7: Triangle count}
\usage{
compute_F7(adj)
}
\arguments{
\item{adj}{Adjacency matrix}
}
\value{
Total number of triangles in the graph
}
\description{
Compute F7: Triangle count
}
