% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Summaries.R
\name{compute_F8}
\alias{compute_F8}
\title{Compute F8: Global clustering coefficient}
\usage{
compute_F8(adj)
}
\arguments{
\item{adj}{Adjacency matrix}
}
\value{
Global transitivity (numeric)
}
\description{
Compute F8: Global clustering coefficient
}
