% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Summaries.R
\name{compute_F9}
\alias{compute_F9}
\title{Compute F9: Negated average path length (reachability metric)}
\usage{
compute_F9(adj)
}
\arguments{
\item{adj}{Adjacency matrix}
}
\value{
Negative of average shortest path length (numeric)
}
\description{
Compute F9: Negated average path length (reachability metric)
}
