% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Summaries.R
\name{compute_scan_stat}
\alias{compute_scan_stat}
\title{Compute scan statistic for a given neighborhood scale}
\usage{
compute_scan_stat(g, k)
}
\arguments{
\item{g}{An igraph graph object}

\item{k}{Integer, neighborhood order (e.g., 1 for 1-hop)}
}
\value{
Maximum number of edges in any k-hop neighborhood (integer)
}
\description{
Compute scan statistic for a given neighborhood scale
}
