% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_dcsbm.R
\name{CP_dcsbm}
\alias{CP_dcsbm}
\title{DCSBM Statistics for Changepoint Detection}
\usage{
CP_dcsbm(A, C = NULL, K = 2)
}
\arguments{
\item{A}{A list of matrices, ordered sequence of adjacency matrices.}

\item{C}{A vector, categorical/integer community memberships for computing local statistics. Defaults to \code{NULL}.}

\item{K}{An integer, number of communities for normalized spectral clustering (based on the first adjacency matrix in the sequence). Defaults to \code{2}. Only used if \code{Z} is unspecified, otherwise \code{K = max(C)}.}
}
\value{
A \code{length(A)}\eqn{\times}\code{(1 + choose(K+1,2) + K)} matrix with named columns, with the summary statistics computed for each adjacency matrix.
}
\description{
Computes 9 graph summary statistics (or graph invariants),
described in \href{https://doi.org/10.1002/qre.2520}{Wilson et al., (2019)}.
for an ordered sequence of undirected networks with no self-loops.
\enumerate{
\item \code{pimax}: proportion of nodes in largest community, according to spectral clustering.
\item \code{Bkl}: local densities for \eqn{1 \leq l \leq k \leq K}.
\item \code{thetak}: sample standard deviations of block \eqn{k} degree-correction parameters.
}
}
\examples{
CP <- list(type='kidneyegg',t_start=5,n_egg=20,p=0.5,q=0.8)
data <- Simulate_CP(100,10,CP=CP)
A <- data$A; C <- nettest:::Z_to_C(data$Z1)
test <- CP_dcsbm(A,C)
}
