% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_laplacian.R
\name{CP_laplacian}
\alias{CP_laplacian}
\title{Graph Laplacian Spectral Anomaly Scores for Changepoint Detection}
\usage{
CP_laplacian(
  A,
  context_method = "svd",
  context_short = 5,
  context_long = 25,
  laplacian = "normalized",
  keep = NULL
)
}
\arguments{
\item{A}{A list of matrices, ordered sequence of adjacency matrices.}

\item{context_method}{Either \code{'svd'} or \code{'mean'}, method for constructing spectral context vectors. Defaults to \code{'svd'}.}

\item{context_short}{Integer, short-term context window size, defaults to \code{5}.}

\item{context_long}{Integer, long-term context window size, defaults to \code{25}.}

\item{laplacian}{Either \code{'normalized'} or \code{'unnormalized'}, which graph Laplacian to use. Defaults to \code{'normalized'}.}

\item{keep}{Integer, how many eigenvalues to keep from Laplacian spectrum.}
}
\value{
A \code{length(A)}\eqn{\times}\code{2} matrix with named columns, containing the 2 anomaly scores.
}
\description{
Computes 2 normalized anomaly scores based on Laplacian spectra,
described in \href{https://doi.org/10.1145/3394486.3403077}{Huang et al., (2020)}.
for an ordered sequence of undirected networks with no self-loops.
\enumerate{
\item \code{Z}: raw anomaly score (cosine distance from short-term or long-term context vectors).
\item \code{Zstar}: change in raw anomaly score.
}
}
\examples{
CP <- list(type='kidneyegg',t_start=30,n_egg=20,p=0.5,q=0.8)
A <- Simulate_CP(40,40,CP=CP)$A
test <- CP_laplacian(A)
}
