% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_shewhart_plot.R
\name{CP_shewhart_plots}
\alias{CP_shewhart_plots}
\title{Plot Time Series with Shewhart-style Control Bands}
\usage{
CP_shewhart_plots(
  series,
  baseline = NULL,
  makeplot = TRUE,
  makeCP = NULL,
  pdfname = NULL
)
}
\arguments{
\item{series}{A matrix, each column contains a univariate time series of length \code{m}.}

\item{baseline}{An integer, sample size of the leading portion used to estimate control limits. Defaults to the minimum of \code{ceiling(m/2)} and \code{25}.}

\item{makeplot}{A Boolean, should a plot be produced. Defaults to \code{TRUE}.}

\item{makeCP}{A numeric vector of changepoints to mark on plots. Defaults to \code{NULL}.}

\item{pdfname}{A string, file location and name to save as a pdf plot. Defaults to \code{NULL}.}
}
\value{
A dataframe containing baseline sample means and standard deviations for each time series.
If \code{makeplot=TRUE}, also produces a plot of Shewhart control charts.
}
\description{
Computes summary statistics and produces plots for Shewhart-style monitoring of time series derived from sequences
of networks. Plots for each time series visualize a sample mean plus/minus 3 sample standard deviations, calibrated
from a baseline window. Additional options to add known changepoints, and save plot output to PDF.
}
