% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_summaries.R
\name{CP_summaries}
\alias{CP_summaries}
\title{Graph Summaries (Invariants) for Changepoint Detection}
\usage{
CP_summaries(A)
}
\arguments{
\item{A}{A list of matrices, ordered sequence of adjacency matrices.}
}
\value{
A \code{length(A)}\eqn{\times}\code{9} matrix with named columns, with the 9 summary statistics computed for each adjacency matrix.
}
\description{
Computes 9 graph summary statistics (or graph invariants), described in \href{https://ieeexplore.ieee.org/document/6380528}{Park et al., (2012)}.
for an ordered sequence of undirected networks with no self-loops.
\enumerate{
\item \code{density}: Edge density, number of edges divided by number of node pairs.
\item \code{maxdeg}: Maximum node degree.
\item \code{maxeig}: Maximum eigenvalue of adjacency matrix.
\item \code{scan-1}: Maximum number of edges over 1-hop neighbourhoods.
\item \code{scan-2}: Maximum number of edges over 2-hop neighbourhoods.
\item \code{scan-3}: Maximum number of edges over 3-hop neighbourhoods.
\item \code{triangle}: Triangle density, number of triangles divided by number of node triples.
\item \code{gclust}: Global clustering coefficient, a measurement of triangle closure.
\item \code{apath}: Negative of average path length over all node pairs. Disconnected node pairs are assigned a path length equal to twice the maximum observed (finite) path lengths.
}
}
\examples{
CP <- list(type='kidneyegg',t_start=5,n_egg=20,p=0.5,q=0.8)
A <- Simulate_netCP(100,10,CP=CP)$A
test <- CP_summaries(A)
}
