% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot_ASE_EPA_test.R
\name{LowRankTests}
\alias{LowRankTests}
\title{LowRankTests}
\usage{
LowRankTests(A, B, r, sig, bs)
}
\arguments{
\item{A}{cell array containing network in 1st population; each cell is a sparse EPAacency matrix}

\item{B}{cell array containing network in 2nd population, each cell being a sparse EPAacency matrix}

\item{r}{scalar specifying rank of population EPAacency}

\item{sig}{significance level for acceptance of null hypothesis}

\item{bs}{bootstrap times}
}
\value{
tASE: output of Boot-ASE test (1 if null is rejected, 0 otherwise)\cr
tEPA: output of Boot-EPA test (1 if null is rejected, 0 otherwise)\cr
pvalASE: p-value for Boot-ASE test\cr
pvalEPA: p-value for Boot-EPA test
}
\description{
This function is used to perform the two-sample test to reveal whether the two set of networks from identical population when m=1
This function performs the Boot-ASE and Boot-EPA tests, which are two-sample testing procedures for large graphs.
This test includes tests which are based on adjacency spectral embedding (ASE) and estimated population adjacency (EPA), respectively. \cr\cr
The test involves computing the rejection decision and the p-value, which is computed to ensure that the null hypothesis (i.e., both populations have the same population adjacency) is rejected only when the test statistic is in the upper α-quantile for both approximate distributions.
}
\examples{
A <- genSparseGraph(1, model1) # sample size, m=1
B <- genSparseGraph(1, model2)
test_result = LowRankTests(A, B, 100, 0.05, 200)


}
