% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_netmodel.R
\name{Simulate_netmodel}
\alias{Simulate_netmodel}
\title{Simulation for (Samples of) Networks}
\usage{
Simulate_netmodel(n, model = list(), m = 1)
}
\arguments{
\item{n}{An integer, number of nodes in each network.}

\item{model}{A named list, provides additional model information:
\describe{
\item{name}{One of \code{'ER'}, \code{'SBM'}, \code{'LSM'}, \code{'wSBM'}, \code{'wLSM'}, describing the model type.
\code{ER}, \code{SBM}, and \code{LSM} produce binary adjacency matrices; \code{wSBM} and \code{wLSM} produce weighted
adjacency matrices with Gaussian edges.}
\item{directed}{A Boolean, are the network edges directed? Defaults to \code{FALSE}.}
\item{self_loops}{A Boolean, are self-loops allowed? Defaults to \code{FALSE}.}
\item{dispersion}{Edge dispersion parameter. For weighted networks, corresponds to edge variance.}
\item{p}{For \code{ER}, overall edge probability.}
\item{Z}{For \code{SBM}, \code{LSM}, \code{wSBM} and \code{wLSM}, prespecified \eqn{n \times K} community membership matrix or \eqn{n \times d} latent position matrix.
If unspecified, community memberships are generated using \code{model$Pi}, and latent space positions are random Gaussian or Dirichlet (for RDPG).}
\item{Y}{For directed \code{LSM} and \code{wLSM}, prespecified \eqn{n \times d} right-hand side latent position matrix.}
\item{Pi}{For \code{SBM} and \code{wSBM}, \eqn{K}-vector of community membership probabilities.}
\item{B}{For \code{SBM} and \code{wSBM}, \eqn{K \times K} matrix of connection probabilties or means.}
\item{theta}{For \code{SBM} and \code{wSBM}, \eqn{n}-vector of degree correction parameters. Defaults to a vector of \code{1}'s.}
\item{d}{For \code{LSM} and \code{wLSM}, integer latent space dimension.}
\item{link}{For \code{LSM}, either \code{'logit'} or \code{'identity'}. If \code{'identity'}, edge parameters are not passed through the \code{expit} function.}
\item{similarity}{For \code{LSM} and \code{wLSM}, either \code{'ip'} or \code{'dist'} corresponding to inner product and Euclidean distance similarity functions.}
\item{alpha}{For \code{LSM} with \code{model$similarity='dist'}, intercept term for edge probabilities.}
}}

\item{m}{An integer, number of repeated samples.}
}
\value{
A list containing:
\item{A}{A matrix or list of \code{m} matrices, simulated adjacency matrix(es).}
\item{Z}{An \eqn{n \times K} community membership matrix or \eqn{n \times d} latent position matrix used to generate each sample.}
\item{Y}{An \eqn{n \times K} right-hand side community membership matrix or \eqn{n \times d} right-hand side latent position matrix used to generate each sample (only if \code{model$directed=TRUE}).}
}
\description{
Simulates an adjacency matrix or list of \code{m} adjacency matrices on \code{n} nodes from one of several named network models.
Additional model information is specified as part of the \code{model} argument, see below.
}
\examples{
set.seed(12)
n <- 40

# (1) ER
model1 <- list(name='ER',p=0.3)
data1 <- Simulate_netmodel(n,model=model1)

# (2) SBM fixed Z
Z2 <- nettest:::C_to_Z(sample(1:2,n,replace=TRUE),2)
B2 <- matrix(c(0.7,0.3,0.3,0.7),2,2)
model2 <- list(name='SBM',Z=Z2,B=B2)
data2 <- Simulate_netmodel(n,model=model2)

# (3) DCSBM random Z
Pi3 <- c(0.5,0.5)
theta3 <- stats::runif(n,min=0.8,1.2)
model3 <- list(name='SBM',B=B2,Pi=Pi3,theta=theta3)
data3 <- Simulate_netmodel(n,model=model3)

# (4) binary RDPG fixed Z
Z4 <- MCMCpack::rdirichlet(n,rep(1,2))
model4 <- list(name='LSM',Z=Z4,link='identity',similarity='ip')
data4 <- Simulate_netmodel(n,model=model4)

# (5) binary LSM random Z (dist sim + logit link)
model5 <- list(name='LSM',d=2,link='logit',similarity='dist',alpha=1)
data5 <- Simulate_netmodel(n,model=model5)

# (6) Gaussian, block constant/SBM
model6 <- list(name='wSBM',B=3*B2,Pi=Pi3)
data6 <- Simulate_netmodel(n,model=model6)

# (7) Gaussian, LSM random Z (ip sim)
model7 <- list(name='wLSM',d=3,dispersion=3,similarity='ip')
data7 <- Simulate_netmodel(n,model7)
}
