% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spectral_clus}
\alias{spectral_clus}
\title{spectral_clus}
\usage{
spectral_clus(C, r)
}
\arguments{
\item{C}{a matrix}

\item{r}{r communities in G}
}
\value{
returns a vector of length n containing the cluster assignment of each node.
}
\description{
This function computes the cluster assignment of each node, serving as an auxiliary functions for the function: Asymp_TW
}
\examples{
C = tcrossprod(1:10)
spectral_clus(C, r = 3)
}
