% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spectral_clust}
\alias{spectral_clust}
\title{(Normalized) Spectral Clustering}
\usage{
spectral_clust(M, d)
}
\arguments{
\item{M}{a square \code{n}\eqn{\times}\code{n} matrix}

\item{d}{number of clusters}
}
\value{
returns a vector of length \code{n} containing the integer-valued cluster assignments.
}
\description{
This function computes the cluster assignment based on the normalized singular vectors of a matrix \code{M}.
}
\examples{
M = tcrossprod(1:10)
spectral_clust(M, d = 3)
}
