% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TW_test.R
\name{TW_test}
\alias{TW_test}
\title{Tracy Widom Test}
\usage{
TW_test(A, B, P_method = "SBA", delta = 0.2)
}
\arguments{
\item{A}{A matrix or list of matrices, adjacency matrix(es) for the first sample.}

\item{B}{A matrix or list of matrices, adjacency matrix(es) for the second sample.}

\item{sig}{significance level for acceptance of null hypothesis}

\item{r}{r communities in G}
}
\value{
acceptance/rejection decision & p-value for a normal distribution
}
\description{
The TW test is a two-sample testing procedure for large graphs.
The TW test is based on the Wishart distribution and uses the Tracy-Widom law to approximate the null distribution of a test statistic.\cr\cr
The TW test then uses the Tracy-Widom law to approximate the null distribution of T_TW under certain assumptions about the underlying distributions.
If the value of T_TW exceeds a certain threshold, then the null hypothesis (i.e., both populations have the same population adjacency) is rejected.
}
\details{
An asymptotic test for network two-sample testing.
It is used to test the global null hypothesis under the undirected IER model with no self-loops.
The test is based on the largest singular value of a suitably scaled difference of mean adjacency matrices.
The test statistic is asymptotically Tracy-Widom distributed with parameter \eqn{\beta=1} as \eqn{n \rightarrow \infty}, as long as \eqn{m} does not grow too fast.
}
\examples{
m1 <- list(name='SBM',B=matrix(0.3,2,2),Pi=c(0.5,0.5))
m2 <- m1; m2$B <- matrix(c(1,0.3,0.3,1),2,2)
A <- Simulate_netmodel(n=40,model=m1,m=4)$A
B <- Simulate_netmodel(n=40,model=m2,m=4)$A
test <- Twosample_tw(A,B,P_method='SBA')
}
