% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twosample_anorm.R
\name{Twosample_anorm}
\alias{Twosample_anorm}
\title{Two-sample Asymptotic Normal Test}
\usage{
Twosample_anorm(A, B)
}
\arguments{
\item{A}{A list of matrices, adjacency matrices for the first sample.}

\item{B}{A list of matrices, adjacency matrices for the second sample.}
}
\value{
A list containing:
\item{p.value}{The (two-sided) \eqn{p}-value for the test}
\item{statistic}{The asymptotically normal test statistic}
}
\description{
An asymptotic test for network two-sample testing, described in \href{https://arxiv.org/abs/1811.12752}{Ghoshdastidar & von Luxburg, (2018)}.
It is used to test the global null hypothesis under the undirected IER model with no self-loops.
The test is based on an estimate of the Frobenius norm of the difference in expected adjacency matrices,
estimated by sample splitting.
The test statistic is asymptotically normal as \eqn{n \rightarrow \infty}.
This test requires at least 2 networks in each sample.
}
\examples{
m1 <- list(name='SBM',B=matrix(0.3,2,2),Pi=c(0.5,0.5))
m2 <- m1; m2$B <- matrix(c(1,0.3,0.3,1),2,2)
A <- Simulate_netmodel(n=10,model=m1,m=4)$A
B <- Simulate_netmodel(n=10,model=m2,m=4)$A
test <- Twosample_anorm(A,B)
}
