% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twosample_boot_ase.R
\name{Twosample_boot_ase}
\alias{Twosample_boot_ase}
\title{Two-sample Bootstrap Tests using ASE}
\usage{
Twosample_boot_ase(A, B, d, nboot = 500, return_null = FALSE)
}
\arguments{
\item{A}{A matrix or list of matrices, adjacency matrix(es) for the first sample.}

\item{B}{A matrix or list of matrices, adjacency matrix(es) for the second sample.}

\item{d}{rank for ASE embedding}

\item{nboot}{number of bootstrap replicates, defaults to \code{500}.}

\item{return_null}{Boolean, should the full bootstrap null distribution be returned? Defaults to \code{FALSE}}
}
\value{
A list containing:
\item{p.value}{Named vector containing the bootstrap \eqn{p}-values for the ASE and EPA tests (with continuity correction)}
\item{statistic}{Named vector containing the test statistics for the ASE and EPA tests}
\item{null_dist}{If \code{return_null=TRUE}, a list of 2 \code{nboot}\eqn{\times}\code{2} matrices containing the bootstrap samples for the ASE and EPA tests}
}
\description{
Two bootstrap tests for network two-sample testing, described in \href{https://arxiv.org/abs/1811.12752}{Ghoshdastidar & von Luxburg, (2018)},
based on methods originally developed in \href{https://doi.org/10.1080/10618600.2016.1193505}{Tang et al., (2017)}.
Both are used to test the global null hypothesis under the undirected IER model with no self-loops, and assuming the
expected adjacency matrices have rank \code{d}.
The ASE test statistic is based on the (Frobenius norm) difference in the aligned adjacency spectral embeddings for
each sample.
The EPA test statistic is based on the (Frobenius norm) difference in the estimated expected adjacency matrices.
The test is calibrated using a parametric bootstrap.
This test supports one network per sample.
}
\examples{
m1 <- list(name='SBM',B=matrix(0.3,2,2),Pi=c(0.5,0.5))
m2 <- m1; m2$B <- matrix(c(1,0.3,0.3,1),2,2)
A <- Simulate_netmodel(n=10,model=m1)$A
B <- Simulate_netmodel(n=10,model=m2)$A
test <- Twosample_boot_ase(A,B,d=2,nboot=40,return_null=TRUE)

}
