% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twosample_boot_shuffle.R
\name{Twosample_boot_shuffle}
\alias{Twosample_boot_shuffle}
\title{Two-sample Bootstrap Shuffling Tests}
\usage{
Twosample_boot_shuffle(A, B, nboot = 500, return_null = FALSE)
}
\arguments{
\item{A}{A list of matrices, adjacency matrices for the first sample.}

\item{B}{A list of matrices, adjacency matrices for the second sample.}

\item{nboot}{number of bootstrap replicates, defaults to \code{500}.}

\item{return_null}{Boolean, should the full bootstrap null distribution be returned? Defaults to \code{FALSE}}
}
\value{
A list containing:
\item{p.value}{Named vector containing the bootstrap \eqn{p}-values for the Frobenius and operator tests (with continnuity correction)}
\item{statistic}{Named vector containing the test statistics for the Frobenius and operator tests}
\item{null_dist}{If \code{return_null=TRUE}, \code{nboot}\eqn{\times}\code{2} matrix containing the bootstrap samples for the Frobenius and operator norm tests}
}
\description{
Two bootstrap tests for network two-sample testing.
Both are used to test the global null hypothesis under the undirected IER model with no self-loops.
The Frobenius test statistic is an estimate of the Frobenius norm of the difference in expected adjacency matrices.
The operator test statistic is an estimate of the operator norm of the difference in expected adjacency matrices.
The test is calibrated using a bootstrap which shuffles the two samples.
This test requires at least 2 networks in each sample.
}
\examples{
m1 <- list(name='SBM',B=matrix(0.3,2,2),Pi=c(0.5,0.5))
m2 <- m1; m2$B <- matrix(c(1,0.3,0.3,1),2,2)
A <- Simulate_netmodel(n=10,model=m1,m=4)$A
B <- Simulate_netmodel(n=10,model=m2,m=4)$A
test <- Twosample_boot_shuffle(A,B,nboot=40,return_null=TRUE)
}
