% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twosample_chisq.R
\name{Twosample_chisq}
\alias{Twosample_chisq}
\title{Two-sample Chi-squared Test}
\usage{
Twosample_chisq(A, B, cov_method = "diag")
}
\arguments{
\item{A}{A list of matrices, adjacency matrices for the first sample.}

\item{B}{A list of matrices, adjacency matrices for the second sample.}

\item{cov_method}{covariance estimation method, one of \code{'diag'}, \code{'shrink'}, or \code{'full'}. Full covariance
will be singular unless there are more total samples than edges. Defaults to \code{'diag'}.}
}
\value{
A list containing:
\item{p.value}{The \eqn{p}-value for the test}
\item{statistic}{The asymptotically chi-square test statistic}
\item{df}{The degrees of freedom of the chi-square null distribution}
}
\description{
An asymptotic test for network two-sample testing.
It is used to test the global null hypothesis under the undirected IER model with no self-loops.
The test is based on the sum of squared difference in edgewise means.
The test statistic is asymptotically chi-squared distributed for fixed \eqn{n}, as \eqn{m \rightarrow \infty} (where \eqn{m} is the number of networks in each sample).
This test requires at least 2 networks in each sample; it will tend to perform poorly for small \eqn{m} or sparse graphs.
}
\examples{
m1 <- list(name='SBM',B=matrix(0.3,2,2),Pi=c(0.5,0.5))
m2 <- m1; m2$B <- matrix(c(1,0.3,0.3,1),2,2)
A <- Simulate_netmodel(n=10,model=m1,m=4)$A
B <- Simulate_netmodel(n=10,model=m2,m=4)$A
test <- Twosample_chisq(A,B,cov_method='shrink')
}
