% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twosample_mesoscale.R
\name{Twosample_mesoscale}
\alias{Twosample_mesoscale}
\title{Two-sample Mesoscale Projection Tests}
\usage{
Twosample_mesoscale(
  A,
  B,
  hyp_set,
  edge_type = "weighted",
  d = 1,
  directed = TRUE,
  self_loops = TRUE,
  proj_type = "impute",
  var_type = "basic",
  centered = FALSE,
  masked_set = list(NULL, NULL)
)
}
\arguments{
\item{A}{A matrix or list of matrices, adjacency matrix(es) for the first sample.}

\item{B}{A matrix or list of matrices, adjacency matrix(es) for the second sample.}

\item{hyp_set}{the hypothesis set of node pairs, specified as one of: (1) a 2-element list
containing the incident row and column indices (rectangle), (2) a list of 2-element lists each containing
incident row and column indices (collection of rectangles), or (3) a 2-column matrix of (row,column) index
pairs (unstructured hypothesis set).}

\item{edge_type}{a string, either \code{'weighted'} or \code{'binary'} depending on the possible
edge values. The test with weighted edges is performed assuming Gaussian edges. Defaults to \code{'weighted'}.}

\item{d}{a positive integer, specifies the dimension of projection in the mesoscale test. Defaults to \code{1}.}

\item{directed}{a Boolean, specifies whether the network is directed. If \code{directed=FALSE}, the
testing methodology will be adjusted to account for symmetry. Defaults to \code{TRUE}.}

\item{self_loops}{a Boolean, if \code{FALSE} the test will ignore diagonal entries. Defaults to \code{TRUE}.}

\item{proj_type}{a string, either \code{'impute'} or \code{'one_step'} to choose the projection learning
approach. Recommendation is \code{'one_step'} for rectangular hypothesis sets, \code{'impute'} for unstructured
hypothesis sets. Defaults to \code{'impute'}.}

\item{var_type}{a string, \code{'basic'}, \code{'orth'}, or \code{'quasi'} to choose the variance estimation approach. For
binary edges, \code{'quasi'} corresponds to fitting with a quasibinomial GLM.
For weighted edges, \code{'orth'} calibrates the variance based on the variability of the edges orthogonal to the projected data, while
\code{'basic'} uses the variability within the projected data. Defaults to \code{'basic'}}

\item{centered}{a Boolean, specifies whether to center the projections to ignore differences in the overall
edge mean on the mesoscale set. Defaults to \code{FALSE}.}

\item{masked_set}{a masked set of node pairs ignored in both the projection learning and testing phases
of the mesoscale methodology. Specified in the same format as \code{hyp_set}. Defaults to the empty set.}
}
\value{
A list containing:
\item{p.value}{The \eqn{p}-value for the test}
\item{statistic}{The asymptotically chi-square test statistic}
\item{df}{The degrees of freedom of the chi-square null distribution}
}
\description{
Asymptotic tests for network two-sample testing, described in \href{https://arxiv.org/abs/2410.17046}{MacDonald et al., (2024)}.
Used to test a mesoscale null hypothesis for a prespecified collection
(hypothesis set) of edges, under an exponential family edge model.
It is compatible with either weighted (Gaussian) edge or binary edge networks
(specified with argument \code{edge_type}), undirected or directed networks
(\code{directed}), and networks with or without self loops (\code{self_loops}).
The mesoscale projection testing methodology uses edge variables outside the hypothesis set
to learn a projection of the edges in the hypothesis set, to reduce the dimension of the
test and improve test power. This test supports one network per sample.
}
\examples{
model <- list(name='SBM',B=matrix(c(0.5,0.3,0.3,0.5),2,2),Pi=c(0.5,0.5),directed=TRUE,self_loops=TRUE)
data <- Simulate_netmodel(n=10,model,m=4,twosample=TRUE)
A <- data$A1; B <- data$A2

# hypothesis set specified as a rectangle
meso1 <- Mesoscale_test(A,B,
                        hyp_set=list(1:4,1:4),
                        edge_type='binary',d=1,directed=TRUE,self_loops=TRUE,
                        proj_type='one_step',
                        var_type='basic')

# same hypothesis set specified as unstructured indices
meso2 <- Mesoscale_test(A,B,
                        hyp_set=cbind(c(1,1,1,2,2,3),c(2,3,4,3,4,4)),
                        edge_type='binary',d=1,directed=TRUE,self_loops=TRUE,
                        proj_type='one_step',
                        var_type='basic')
# check that meso1 and meso2 give the same result

# hypothesis set specified as two rectangles, test with centering
meso3 <- Mesoscale_test(A,B,
                        hyp_set=list(list(1:2,1:2),list(9:10,9:10)),
                        edge_type='binary',d=1,directed=TRUE,self_loops=FALSE,
                        proj_type='impute',
                        var_type='quasi',
                        centered=TRUE)

}
