% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twosample_thirdpower.R
\name{Twosample_thirdpower}
\alias{Twosample_thirdpower}
\title{Two-sample 3rd Power Test}
\usage{
Twosample_thirdpower(A, B, P_method = "SBA", delta = 0.5, d = 2, ndiag = 100)
}
\arguments{
\item{A}{A matrix or list of matrices, adjacency matrix(es) for the first sample.}

\item{B}{A matrix or list of matrices, adjacency matrix(es) for the second sample.}

\item{P_method}{P-matrix estimation method, one of \code{'NBS'}, \code{'SBA'}, \code{'USVT'} or \code{spectral_clust}.
All but \code{spectral_clust} are implemented using the \code{graphon} package.
Defaults to \code{'SBA'}. Note that \code{'NBS'} and \code{'USVT'} can perform erratically when \eqn{n} is very small.}

\item{delta}{precision parameter larger than zero, only used if \code{P_method='SBA'}. Defaults to \code{0.5}.}

\item{d}{number of clusters, only used if \code{P_method='spectral_clust'}. Defaults to \code{2}.}

\item{ndiag}{number of random replications of the diagonal adjustment. Defaults to \code{100}.}
}
\value{
A list containing:
\item{p.value}{The (two-sided) \eqn{p}-value for the test}
\item{statistic}{The asymptotically normal test statistic, averaged over random replications of the diagonal adjustment.}
}
\description{
An asymptotic test for network two-sample testing, described in \href{https://www.jstor.org/stable/27280115}{Chen et al., (2024)}.
It is used to test the global null hypothesis under the undirected IER model with no self-loops.
The test is based on the trace of the 3rd power of a suitably scaled difference of mean adjacency matrices, with a random adjustment to the diagonal entries.
The test statistic is asymptotically normal as \eqn{n \rightarrow \infty}, as long as \eqn{m} does not grow too fast.
}
\examples{
m1 <- list(name='SBM',B=matrix(0.3,2,2),Pi=c(0.5,0.5))
m2 <- m1; m2$B <- matrix(c(1,0.3,0.3,1),2,2)
A <- Simulate_netmodel(n=40,model=m1,m=4)$A
B <- Simulate_netmodel(n=40,model=m2,m=4)$A
test <- Twosample_thirdpower(A,B,P_method='SBA')
}
