% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twosample_tw.R
\name{Twosample_tw}
\alias{Twosample_tw}
\title{Two-Sample Tracy-Widom Test}
\usage{
Twosample_tw(A, B, P_method = "SBA", delta = 0.2)
}
\arguments{
\item{A}{A matrix or list of matrices, adjacency matrix(es) for the first sample.}

\item{B}{A matrix or list of matrices, adjacency matrix(es) for the second sample.}

\item{P_method}{P-matrix estimation method, one of \code{'NBS'}, \code{'SBA'}or \code{'USVT'}, implemented using the \code{graphon} package.
\code{SBA} requires an additional tuning parameter \code{delta}. Defaults to \code{'SBA'}. Note that \code{'NBS'} and \code{'USVT'} can perform erratically when \eqn{n} is very small.}

\item{delta}{precision parameter larger than zero, only used if \code{P_method='SBA'}. Defaults to \code{0.2}.}
}
\value{
A list containing:
\item{p.value}{The (one-sided) \eqn{p}-value for the test}
\item{statistic}{The asymptotically \eqn{TW(1)} test statistic.}
}
\description{
An asymptotic test for network two-sample testing.
It is used to test the global null hypothesis under the undirected IER model with no self-loops.
The test is based on the largest singular value of a suitably scaled difference of mean adjacency matrices.
The test statistic is asymptotically Tracy-Widom distributed with parameter \eqn{\beta=1} (denoted \eqn{TW(1)}) as \eqn{n \rightarrow \infty}.
}
\examples{
m1 <- list(name='SBM',B=matrix(0.3,2,2),Pi=c(0.5,0.5))
m2 <- m1; m2$B <- matrix(c(1,0.3,0.3,1),2,2)
A <- Simulate_netmodel(n=40,model=m1,m=4)$A
B <- Simulate_netmodel(n=40,model=m2,m=4)$A
test <- Twosample_tw(A,B,P_method='SBA')
}
