% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymp_normal_test.R
\name{Asymp_normal}
\alias{Asymp_normal}
\title{Asymptotic Normal Test}
\usage{
Asymp_normal(A, B, sig)
}
\arguments{
\item{A}{cell array containing networks in 1st population; each cell is a sparse adjacency matrix}

\item{B}{cell array containing networks in 2nd population, each cell being a sparse adjacency matrix}

\item{sig}{significance level for acceptance of null hypothesis}
}
\value{
acceptance/rejection decision & p-value for a normal distribution
}
\description{
The Asymp-Normal test is an asymptotic test for two-sample testing of large graphs.
It is used to determine whether two graphs (or graph populations) have the same population adjacency.\cr\cr
The test involves computing rejection decision and the p-value, which is the probability that the null hypothesis (i.e., both graphs have the same population adjacency) is true.
}
\examples{
simu = twosamp_twoblock(n = 4, m = 4, p = 0.3, q = 0.3, epsilon = 0.7)
test_result = Asymp_normal(simu$A_G, simu$A_H, 0.05)
}
