% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot_ASE_EPA_test.R
\name{Boot_ASE_computeStat}
\alias{Boot_ASE_computeStat}
\title{Boot-ASE/EPA Compute Statistics}
\usage{
Boot_ASE_computeStat(A1, B1, r)
}
\arguments{
\item{A1}{a sparse EPAacency matrix}

\item{B1}{a sparse EPAacency matrix}

\item{r}{scalar specifying rank of population EPAacency}
}
\value{
ASE statistic and EPA statistic
}
\description{
Boot-ASE/EPA Compute Statistics
}
\details{
This function computes the test statistics for Boot-ASE/EPA test, serving as an auxiliary functions for the function: LowRankTests
}
\examples{
simu = twosamp_twoblock(n = 4, m = 4, p = 0.3, q = 0.3, epsilon = 0.7)
Boot_ASE_computeStat(simu$A_G[[1]], simu$A_H[[1]], 100)


}
