% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot_spectral_test.R
\name{Boot_Spectral_computeStat}
\alias{Boot_Spectral_computeStat}
\title{Boot-Spectral Compute Statistics}
\usage{
Boot_Spectral_computeStat(A1, B1)
}
\arguments{
\item{A1}{a sparse adjacency matrix}

\item{B1}{a sparse adjacency matrix}
}
\value{
\eqn{\Tau_fro} & \eqn{\Tau_spec}
}
\description{
This function computes the test statistics for Boot-Spectral test, serving as an auxiliary functions for the function: ShufflingTests
}
\examples{
A <- genSparseGraph(4, model=list(name='ER',n=10,p=.1))
B <- genSparseGraph(4, model=list(name='ER',n=10,p=.5))
testStat <- Boot_Spectral_computeStat(A, B)



}
