% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad_plot_utils.R
\name{compute_z_scores}
\alias{compute_z_scores}
\title{Compute LAD Z and Z* Scores}
\usage{
compute_z_scores(spectra, s = 5, l = 10)
}
\arguments{
\item{spectra}{List of normalized Laplacian spectra (from each time step)}

\item{s}{Size of the short-term window (default: 5)}

\item{l}{Size of the long-term window (default: 10)}
}
\value{
A list with components:
\item{Z}{Raw anomaly scores (cosine dissimilarity)}
\item{Z_star}{Change-in-score anomaly scores}
}
\description{
Compute LAD Z and Z* Scores
}
