% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad_plot_utils.R
\name{get_context_vector}
\alias{get_context_vector}
\title{Compute Context Vector from Spectra}
\usage{
get_context_vector(spectra_list)
}
\arguments{
\item{spectra_list}{List of normalized singular spectra (each a numeric vector)}
}
\value{
The first left singular vector of the stacked spectra matrix
}
\description{
Compute Context Vector from Spectra
}
