% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad_plot_utils.R
\name{plot_lad_scores}
\alias{plot_lad_scores}
\title{Plot LAD Z and Z* Scores Over Time}
\usage{
plot_lad_scores(
  Z,
  Z_star,
  changepoints = NULL,
  title = "LAD Scores",
  skip_t = NULL
)
}
\arguments{
\item{Z}{Vector of raw anomaly scores (Z)}

\item{Z_star}{Vector of Z* (difference in scores)}

\item{changepoints}{Optional vector of changepoint indices to annotate}

\item{title}{Optional plot title}
}
\value{
No return value; generates plots
}
\description{
Plot LAD Z and Z* Scores Over Time
}
