% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TW_test.R
\name{spectral_clus}
\alias{spectral_clus}
\title{spectral_clus}
\usage{
spectral_clus(A, B, r)
}
\arguments{
\item{A}{a sparse EPAacency matrix}

\item{B}{a sparse EPAacency matrix}

\item{r}{r communities in G}
}
\value{
returns a vector of length n containing the cluster assignment of each node.
}
\description{
This function computes the cluster assignment of each node, serving as an auxiliary functions for the function: Asymp_TW
}
\examples{
spectral_clus(simu$A_G, simu$A_H, r = 3)
}
